+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 - Retry when putting comments
+- Add ability to get email by id 
 - Add ability to upload attachment content for comments
 
 # v0.29.0